requirements.txt
webdriver_manager==3.2.2
cx_Freeze==6.3
pytesseract==0.3.6
selenium==3.141.0
requests==2.24.0
Pillow==8.0.1
PyAutoGUI==0.9.52
PyAutoIt==0.6.0